:recycle: refactor: 将SafeAreaView替换为View，并使用SafeAreaInsets在选项卡之间进行一致的填充
