+++ components/layout/tab-background.tsx
@@ -3,15 +3,20 @@ import { LinearGradient } from "expo-linear-gradient";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
 
 import { useColorScheme } from "@/hooks/use-color-scheme";
-import { DEFAULT_SCHEDULE_VISUAL, useScheduleStore } from "@/store/schedule";
+import { useScheduleStore } from "@/store/schedule";
 
 /**
- * 首页/功能页的全屏背景图。
- * 固定使用默认透明度与模糊值，不受“课表外观”里的滑杆设置影响；
- * 顶部叠一层向下消隐的底色渐变，保证状态栏与页面标题区域清晰。
+ * Tabs shared background. Mounted above the tab navigator so switching tabs
+ * does not remount/reload the image layer.
  */
 export function TabBackground() {
   const backgroundImageUri = useScheduleStore((s) => s.backgroundImageUri);
+  const backgroundImageOpacity = useScheduleStore(
+    (s) => s.backgroundImageOpacity,
+  );
+  const backgroundImageBlurRadius = useScheduleStore(
+    (s) => s.backgroundImageBlurRadius,
+  );
   const isDark = useColorScheme() === "dark";
   const insets = useSafeAreaInsets();
 
@@ -22,6 +27,7 @@ export function TabBackground() {
   return (
     <>
       <Image
+        pointerEvents="none"
         source={{ uri: backgroundImageUri }}
         style={{
           position: "absolute",
@@ -29,10 +35,10 @@ export function TabBackground() {
           left: 0,
           right: 0,
           bottom: 0,
-          opacity: DEFAULT_SCHEDULE_VISUAL.backgroundImageOpacity,
+          opacity: backgroundImageOpacity,
         }}
         contentFit="cover"
-        blurRadius={DEFAULT_SCHEDULE_VISUAL.backgroundImageBlurRadius}
+        blurRadius={backgroundImageBlurRadius}
       />
       <LinearGradient
         colors={[
